New LibScout configuration file
This commit adds a new LibScout configuration file that uses the
TOML format. It's default location is config/LibScout.toml but can be
changed via the -c command line switch.

For now it includes the following operational options (more to come):
- path to log4j config file logback.xml (it's default location changed
  from logging to config dir)
- path to android-sdk.jar. if set, the -a switch is no longer required
  on command line.

This fixes #9 as LibScout can now run from any directory if the config
file is provided via command line.