PackageTree rendering with ascii and unicode
Added new config option packageTree.ascii_rendering that
changes the charset used for rendering the package tree.

If the default (false) is not working, e.g. because the console
cannot properly render the box-drawing unicode characters, setting
this option to true forces the use of ascii characters (a bit
less visually appealing)